fixed the lists dates and added animations on delete

+++ client/src/app/views/codes/activation-codes/activation-codes.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { CodesService } from '../../../services/codes.service';
+import Swal from "sweetalert2";
 
 @Component({
   selector: 'app-activation-codes',
@@ -14,15 +15,37 @@ export class ActivationCodesComponent implements OnInit {
 
   getActivationCodes(): void {
     this.codesService.getActivationCodes().subscribe((res) => {
-      console.log(res)
       this.activationCodes = res
+      this.activationCodes.map(code => {
+        code['createdOn'] = new Date(code['createdOn']).toDateString()
+      })
     })
   }
 
   deleteCode(codeId): void {
-    this.codesService.deleteCode(codeId, 'activation').subscribe((res) => {
-      this.getActivationCodes();
+
+    Swal.fire({
+      title: 'Vous êtes sûr?',
+      text: "Les données supprimées seront perdus",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      cancelButtonText: 'Annuler',
+      confirmButtonText: 'Oui, supprimer!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire(
+          'Supprimé!',
+          'Suppression terminée avec succès',
+          'success'
+        )
+        this.codesService.deleteCode(codeId, 'activation').subscribe((res) => {
+          this.getActivationCodes();
+        })
+      }
     })
+
   }
 
   ngOnInit(): void {
