+++ client/src/app/views/codes/courses-codes/courses-codes.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { CodesService } from '../../../services/codes.service';
+import Swal from "sweetalert2";
 
 @Component({
   selector: 'app-courses-codes',
@@ -14,14 +15,34 @@ export class CoursesCodesComponent implements OnInit {
 
   getCoursesCodes(): void {
     this.codesService.getCoursesCodes().subscribe((res) => {
-      console.log(res)
       this.coursesCodes = res
+      this.coursesCodes.map(code => {
+        code['createdOn'] = new Date(code['createdOn']).toDateString()
+      })
     })
   }
 
   deleteCode(codeId): void {
-    this.codesService.deleteCode(codeId, 'course').subscribe((res) => {
-      this.getCoursesCodes();
+    Swal.fire({
+      title: 'Vous êtes sûr?',
+      text: "Les données supprimées seront perdus",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      cancelButtonText: 'Annuler',
+      confirmButtonText: 'Oui, supprimer!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire(
+          'Supprimé!',
+          'Suppression terminée avec succès',
+          'success'
+        )
+        this.codesService.deleteCode(codeId, 'course').subscribe((res) => {
+          this.getCoursesCodes();
+        })
+      }
     })
   }
